+++ CHANGELOG.md
@@ -10,6 +10,8 @@ Changelog](http://keepachangelog.com/en/1.0.0/).
 
 - TRT-627 - Implemented workflow to retrieve all Harmony services from CMR GraphQL
   along with all associated collections.
+- TRT-630 - Implemented test suite for HyBIG.
+- TRT-628 - Implemented scaffolding to invoke all defined test suites.
 
 TODO